Break up work to reset to factory defaults
Resetting ghilbert to factory defaults used to be done in one shot
in a single POST operation.  However, in the actual app engine
server, this took too long and the operation was terminated
before it completed.  This is a first attempt to break up the
work into chunks small enough that the operation will not time out.



git-svn-id: https://ghilbert.googlecode.com/svn/branches/dlk_datastore@54 c014309e-2312-11df-8901-e16673284e38