Basic support for editing or deleting existing theorems
Using the 'replace' button on the edit page, one can replace
an existing theorem with zero or more thm, defthm, var, tvar, or
kindbind commands. This is a 'dangerous' operation, in that
you are allowed to do it even if the replacement commands don't
verify, or cause later theorems to fail. We probably will want
to add some checks later to make this a bit safer.  However,
I think that some sorts of edits will be difficult to do without
allowing at least temporary breakage of verification.

At present reporting of errors if there are verification failures
as a result of 'replace' or 'save new' is not very good. You can
however look at the auto-generated interface file for a proof
file that you edit; it will show any commands in the proof file
that failed, commented out.



git-svn-id: https://ghilbert.googlecode.com/svn/branches/dlk_datastore@58 c014309e-2312-11df-8901-e16673284e38