Import theorem database (upto 999) from orcat.org repository
The original plan was that the database would contain the authoritative
repository of all theorems. However, to do that right would mean adding
many version control features to the database, and processes for things
like backups.

For expedience, we're going to maintain the authoritative database of
theorems in git, and use the database as a working space. From time to
time, we'll do a git

This commit contains a development of recursive functions (with lots of
supporting infrastructure including ordered pairs, primes, factorials,
gcd, Godel's beta function, and finite sequences). These proofs were
done (as far as I know entirely) by Adam Bliss.