Merge Ghilbert app changes from abliss branch
This commit merges in the changes made on the abliss branch. There are
quite a few small changes (and cleanups to the code), but the big ones
include:

+ Switching from canvas-based editor to textarea

+ Adding proof step suggestion buttons (inference, deduction, unified)

+ Save draft

There's quite a bit of other stuff in the abliss branch (for the orcat
UI), which is not included here - it's for the direct mode changes only.