Start implementing 'showthm' functionality
This is the start of a feature to display a theorem, eventually in nicely
formatted HTML, but as of this commit it just picks the theorem out of the
repo + db stream and outputs it. That might be useful by itself some time,
partly motivating this snapshot.

The implementation is similar to the save functionality, in that it runs the
entire proof file through the verifier (up to and including the theorem to
show). The actual formatting of the theorem will be done by subclassing the
runner, much in parallel with the way the JS implementation subclasses the
thm command in its runner.