Basic showthm functionality
While still rough in a lot of ways, this patch adds a working showthm function,
including Unicode typesetting of the intermediate proof stack results. This
latter is done using the existing typeset.js, primarily to avoid having to
maintain multiple versions of the typesetting codebase. Doing the typesetting
server-side might be a better user experience.