Improve showthm display
This patch improves the showthm display in a number of ways, primarily adding
links to theorems within the same stream.

The interface between scanner proof runner is also made more powerful and
general, with notification of whitespace, newlines, and tokens. In addition,the formatting of the proof output is broken out into a separate class, to
make it cleaner.

It also adds a 404 page when the theorem is not found.