Add /listthms page listing all thms.
This patch adds a page listing all thms in the stream (note that it doesn't
include defthm, which is probably an oversight as those are interesting too).
It also tries to extract a description line from comments preceding the thm,
but in the current state of the repo it's not hugely useful. Adding at least a
one line description to each thm is a great goal.

Also note that I expect the url for this and /showthm/* to change, as it really
should include a module name.