Beginnings of git server support
A skeleton server that can export a git repository through the "dumb" http
protocol, over App Engine. This was tested both with a real repo (you have
symlink .git to _git, though, to avoid permissions problems) and with a
purely synthetic one.

This patch also improves header handling support, so sha headers match, and
also tweaks the content type header setting to avoid duplicate headers in the
result (this should probably be applied to the main branch irrespective).