Complete move to HTML typesetting
This change replaces the canvas-based display of the proof stack with an
HTML version. Thus, typesetting that uses HTML (such as exponentiation)
now works in the proof stack view as well.

At this point, it would make sense to delete all the canvas stuff, but
that's a cleanup left for later.