Add a stage (like the git index)
This change adds a "stage" in the datastore, so that saves get written
into the stage, then commited as a separate step.

A few flaws remain: timezone handling is quite wonky. Also, the new
interfaces in the store object are only implemented in the appengine
store - they should be implemented in the in-memory store too. And
there's no memcache of the stage. There should probably be a ref in the
stage object (corresponding to HEAD in git). Lastly, there's no
transactions or any protection against concurrent writes clobbering each
other.