Basic support for a wiki
This patch starts adding a wiki to the web app. Wiki pages are stored in
the git repo, and a variant of ghmarkup is used to render the content.

Formatting and styling is still very raw, and there's little integration
with the theorems (you can link to them though), but it's a start.