Protect access to git repo by login
Any save action into the git repo requires a login with valid write
permissions. In addition, it uses the identity from the user profile for
the commit message.

The flow is still rough (for example, the login action should redirect
you to the page you were attempting), but at least this version should
be strong enough to push to production.