Add invite code to account creation
I've opened up a mechanism to hash-check an invite code, which gives write
permissions the new account on creation. In fact, this is the only way to
create permissions other than using the admin console to edit the db.

Other minor improvements in signup flow, like offering a helpful random
password, and signing the user in on account creation.

Housekeeping: slightly updated README.