Serve proof editing page from git
With this patch, the theorem edit page is served from the git repository,
rather than from the mix of static and db as before.

Still not ready for prime time, as it can't save to git, and there's
other stuff that doesn't work (drafts, handling of absolute urls in
interface names). But it's a good start.

Also a number of other small cleanups, including getting rid of tabs
from all Python code.