Add git-based thm save
This patch basically duplicates the old save functionality, but splicing
the thm into the file and saving it in git.

This change doesn't affect the JavasSript at all. A number of
improvements are needed, including plumbing the url (not just the thm
name) through the JavaScript client, as the url is currently hardcoded,
so it can't be used to save edits from multiple .gh files. Also, error
reporting needs to be better.