Better error handling in showthm
This patch moves towards being able to handle proof files with errors.
The listthm display shows a red dot on proofs with errors, but
continues. Also, the edit save handler reports the error as a json
response.

Similar changes will need to be made to js to handle proof files with
errors similarly, but not yet done.