Make js verify suppress errors
This patch adds the capability of suppressing errors in the JavaScript
verifier, which allows saving drafts.

It tries to only accept errors within a proof file (ie the types of
errors that would let verify.py -c continue).

Also uses JSON for the save return value - this is more flexible and can
easily be expanded.