Support multi_ack_detailed in git server
This patch upgrades the smart http git protocol to support the
multi_ack_detailed option. The server now parses the "have" commands
from the client, reports the commits that are in common, and only sends
the commits that are not needed (although it currently sends the entire
tree, so the diff could definitely be further optimized).