New theorems for division, exponentiation, modulo
A bunch of work (done from a local repo) in this commit.

Definition of division with some basic properties. Also monotonicity
of exponentiation, and more theorems for minimum and modulo, in support
of the division defintion. And some reorganization (especially trying
to get theorems for le and lt less scattered).