Create workspace patch and autosave
We can now create a patch for the current workspace state (relative to
the base loaded from the server), and autosave that, with a fixed
timeout. It's not yet loading these patches when reconstructing the
workspace state - that's next.

Neil Fraser's diff-match-patch is used as the patching engine (and this
patch checks in the minified JS version).