Switch from diffs to deltas
We now use deltas (computed with own very fast algorithm) instead of
diffs to represent the changes in the workspace relative to the base
version.

Line-based deltas are more robust than character-indexed diffs because
of the different encodings used (utf-16 on the client, utf-8 on the
server). Deltas are also more concise than diffs, especially for block
moves. The algorithm in delta.js is extremely fast.