Use correct symbol context for export
The export_match() function was using the symbol map from the verify
context, but it should have used the symbol map from the statement being
exported. Most of the time, these are the same, but when directly
reexporting a statement that has been imported from another interface,
they may not. Also adds test case for this.

Thanks to Jim Kingdon for the bug report and test case.