Wire up commit from workspace
It's now possible to commit the contents of the workspace, which means
the workspace can now be used for "real work".

The biggest problem is handling of concurrent commits, those made since
the workspace was checked out. Currently it just signals an error. A
better experience would be to commit to a branch, so that you could use
another git client (for example, commandline) to merge. An even better
experience, of course, would be merge tools in the workspace, but
that's a "big-ticket item" and not essential for now.