Adds more tests.
Also fixes a few small bugs in the js and py verifiers which shook out in the
process. Nothing unsound, just robustness checks.

I used 'istanbul' to run a code-coverage analysis of verify.js, and used the
uncovered code as a guide to write more test cases.  With these extra tests, the
tool now reports coverage of 1006 statements and 591 branches. This is up from
810 statements and 423 branches as of commit 6313b24.

I do not think any more coverage of verify.js is possible without significant
refactoring of the file or an expansion of the testsuite capabilities. The
remaining uncovered code is mostly styling stuff.

Note that these tests are only conceived to cover code which is already written,
and thus do not address any "oh, I forgot to check for that!"-type errors.