Improve unification
Fixes a bug in unification, and also does more careful validation of
the graph afterwards.

This is most of the way towards complete proving of the fragment
lacking bound variables, but it still ignores result lines (and thus
won't pass proofs that rely on result lines to fill in terms that
can't otherwise be synthesized).