Start to add infix syntax
Add Pratt parsing of infix operators with precedence.

The infix operators are specified with a new "syntax" block, see the
demo files to see what it looks like.

Not complete, we need prefix and binder syntax at least, not to
mention more complex examples such as [A/x] ph and {A, B}.