Reconstruct expressions at each proof step
This patch contains code that reconstructs the expression at each
proof step, and displays it in the info pane.

The code first reconstructs from the unification graph to an Expr,
then rewrites the Expr to a ParseNode (filling in the actual name of
variables). Then, it typesets those and inserts them into the json.
Lastly, the JS displays that expression in the info pane, and queues
the request to MathJax to typeset it.