Migrate from node indices to step indices
We'll need to be able to query deeper information from a step
(especially to reconstruct the substitution). We were using the node
index as the interface from the unification graph through to the proof
listener, and then the proof listener was converting it to a step
index. This patch changes the interface to steps, and guarantees that
there is a unique step index for each parse node representing a step.